requirements.txt
python-telegram-bot==20.7
openai==1.30.1
transformers==4.35.0
torch==2.3.1
fastapi==0.103.0
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydub==0.25.1
requests==2.31.0
python-multipart==0.0.9
